+++ fastly/ngwaf/v1/signals/doc.go
@@ -1,0 +1,2 @@
+// Package signals contains CRUD operations for signals on the Fastly Next-Gen WAF.
+package signals